refactor: split ontology mixin into CRUD, scoring, and edges
OntologyMixin was 1487 lines with 29 methods across 6 concerns.
Split into three focused modules:

- ontology.py (488 lines, 11 methods): Node CRUD and lifecycle
- ontology_scoring.py (661 lines, 10 methods): Scoring, analytics,
  annealing mutations, and batch operations
- ontology_edges.py (380 lines, 8 methods): Inter-ontology edges
  and proposal execution primitives

Class constants (VALID_ONTOLOGY_EDGE_TYPES, VALID_ONTOLOGY_EDGE_SOURCES)
moved to ontology_edges.py where they're used for edge validation.
