+++ api/app/lib/age_client/__init__.py
@@ -18,17 +18,29 @@ from .base import BaseMixin
 from .ingestion import IngestionMixin
 from .query import QueryMixin
 from .ontology import OntologyMixin
+from .ontology_scoring import OntologyScoringMixin
+from .ontology_edges import OntologyEdgesMixin
 from .vocabulary import VocabularyMixin
 
 
-class AGEClient(VocabularyMixin, OntologyMixin, QueryMixin, IngestionMixin, BaseMixin):
+class AGEClient(
+    VocabularyMixin,
+    OntologyEdgesMixin,
+    OntologyScoringMixin,
+    OntologyMixin,
+    QueryMixin,
+    IngestionMixin,
+    BaseMixin,
+):
     """Client for interacting with Apache AGE knowledge graph database.
 
     Composed from domain mixins:
     - BaseMixin: Connection pool, Cypher execution, AGE type parsing
     - IngestionMixin: Source/Concept/Instance CRUD, graph linking, document metadata
     - QueryMixin: Vector search, learned knowledge, grounding strength, query facade
-    - OntologyMixin: Ontology CRUD, lifecycle, scoring, annealing, inter-ontology edges
+    - OntologyMixin: Ontology node CRUD and lifecycle management
+    - OntologyScoringMixin: Ontology scoring, analytics, and annealing mutations
+    - OntologyEdgesMixin: Inter-ontology edges and proposal execution primitives
     - VocabularyMixin: Relationship type config, CRUD, sync, merge, embeddings
     """
     pass
